requirements.txt
alembic==1.10.3
certifi==2022.12.7
cffi==1.15.1
charset_normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
flask==2.2.3
flask_Migrate==4.0.4
flask_SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
mako==1.2.4
markupSafe==2.1.2
psycopg2-binary==2.9.6
pycparser==2.21
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
sqlalchemy==2.0.9
sqlalchemy-utils==0.37.9
python-dotenv==1.0.0
psycopg2==23.0.1

